move auth functions to auth-common
