+++ enterprise/frontend/src/embedding/auth-common/saml-token-storage.ts
