+++ enterprise/frontend/src/embedding/auth-common/saml.ts
